backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
pydantic
python-dotenv
